qint to quint and further implementations in the ui

+++ src/protocols/tcp.cpp
@@ -2,14 +2,14 @@
 #include "headerAttribute.h"
 
 
-void TCP::initHeader(const IPAddress &srcAdress, const IPAddress &destAdress, const Port &sourcePort, const Port &destinationPort, qint32 seqNumber, qint32 ackNumber, bool ack,bool rst, bool syn, bool fin, qint16 window, Package& data, qint16 dataLength){
+void TCP::initHeader(const IPAddress &srcAdress, const IPAddress &destAdress, const Port &sourcePort, const Port &destinationPort, quint32 seqNumber, quint32 ackNumber, bool ack,bool rst, bool syn, bool fin, quint16 window, Package& data, quint16 dataLength){
     HeaderAttribute srcPort("Source Port",16,sourcePort.getPortNumber());
     HeaderAttribute dstPort("Destination Port",16,destinationPort.getPortNumber());
     HeaderAttribute sequenceNumber("Sequence number",32,seqNumber);
     HeaderAttribute acknowledgementNumber("Acknowledgment number",32,ackNumber);
 
     //Sets the flags for the TCP Header Urgent and Push flag are always 0
-    qint16 flags = 0x0000000000000000;
+    quint16 flags = 0x0000000000000000;
     setFlag(&flags,ack,11);
     setFlag(&flags,rst,13);
     setFlag(&flags,syn,14);
@@ -31,9 +31,9 @@ void TCP::initHeader(const IPAddress &srcAdress, const IPAddress &destAdress, co
                                             dataLength));
 
     //The urgent pointer is always 0 in our case
-    qint16 urgent_pointer = 0b0000000000000000;
+    quint16 urgent_pointer = 0b0000000000000000;
     HeaderAttribute urgentPointer("Urgent Pointer",16,urgent_pointer);
-    HeaderAttribute options("Options",0,0);
+    HeaderAttribute options("Options",static_cast<quint8>(0),static_cast<quint8>(0));
 
     Header* tcpHeader = new Header();
     tcpHeader->setHeaderType(HeaderType::TCP);
@@ -51,7 +51,7 @@ void TCP::initHeader(const IPAddress &srcAdress, const IPAddress &destAdress, co
 }
 
 
-void TCP::setFlag(qint16* flags, bool set, qint16 position){
+void TCP::setFlag(quint16* flags, bool set, quint16 position){
     if (set) {
         // sets the bit at position
         *flags |= (1 << position);
@@ -61,17 +61,17 @@ void TCP::setFlag(qint16* flags, bool set, qint16 position){
     }
 }
 
-qint16 TCP::getTCPChecksum(qint8* sourceAddress, qint8* destinationAddress, qint16 sourcePort, qint16 destinationPort, qint32 seqNumber, qint32 ackNumber, qint16 flags,const char* data, qint16 dataLength) {
+quint16 TCP::getTCPChecksum(quint8* sourceAddress, quint8* destinationAddress, quint16 sourcePort, quint16 destinationPort, quint32 seqNumber, quint32 ackNumber, quint16 flags,const char* data, quint16 dataLength) {
     // TCP Pseudo Header
-    qint32 pseudoHeader = 0;
+    quint32 pseudoHeader = 0;
 
     // Splitting the source IP-Address into 2 16-bit parts with correct byte order
-    qint16 sourceIPAdressPart1 = (sourceAddress[0] << 8) | (sourceAddress[1] & 0xFF);
-    qint16 sourceIPAdressPart2 = (sourceAddress[2] << 8) | (sourceAddress[3] & 0xFF);
+    quint16 sourceIPAdressPart1 = (sourceAddress[0] << 8) | (sourceAddress[1] & 0xFF);
+    quint16 sourceIPAdressPart2 = (sourceAddress[2] << 8) | (sourceAddress[3] & 0xFF);
 
     // Splitting the destination IP-Address into 2 16-bit parts with correct byte order
-    qint16 destinationIPAdressPart1 = (destinationAddress[0] << 8) | (destinationAddress[1] & 0xFF);
-    qint16 destinationIPAdressPart2 = (destinationAddress[2] << 8) | (destinationAddress[3] & 0xFF);
+    quint16 destinationIPAdressPart1 = (destinationAddress[0] << 8) | (destinationAddress[1] & 0xFF);
+    quint16 destinationIPAdressPart2 = (destinationAddress[2] << 8) | (destinationAddress[3] & 0xFF);
 
     pseudoHeader += sourceIPAdressPart1;
     pseudoHeader += sourceIPAdressPart2;
@@ -80,7 +80,7 @@ qint16 TCP::getTCPChecksum(qint8* sourceAddress, qint8* destinationAddress, qint
     pseudoHeader += 6; // IP protocol field for TCP is 6
     pseudoHeader += dataLength;
 
-    qint16 checksum = 0;
+    quint16 checksum = 0;
     //Adding the pseudo Header to the checksum and handling the possible overflow
     while (pseudoHeader >> 16){
         pseudoHeader = (pseudoHeader & 0xFFFF) + (pseudoHeader >> 16);
@@ -96,7 +96,7 @@ qint16 TCP::getTCPChecksum(qint8* sourceAddress, qint8* destinationAddress, qint
 
     //Adding the TCP Data
     for (int i = 0; i < dataLength; i += 2) {
-        qint16 dataWord = (data[i] << 8) | (data[i + 1] & 0xFF);
+        quint16 dataWord = (data[i] << 8) | (data[i + 1] & 0xFF);
         checksum += dataWord;
         if (checksum >> 16)
             checksum = (checksum & 0xFFFF) + (checksum >> 16);
